Move severity to level conversion.
